update host server and common metadata. update several host controllers api.

+++ src/common/mapstr/mapstr.go
@@ -88,7 +88,7 @@ func (cli MapStr) BaseType(key string) (*basetype.Type, error) {
 }
 
 // Int return the value by the key
-func (cli MapStr) Int(key string) (int, error) {
+func (cli MapStr) Int64(key string) (int64, error) {
 
 	switch t := cli[key].(type) {
 	default:
@@ -98,18 +98,18 @@ func (cli MapStr) Int(key string) (int, error) {
 	case int:
 		return t, nil
 	case int16:
-		return int(t), nil
+		return int64(t), nil
 	case int32:
-		return int(t), nil
+		return int64(t), nil
 	case int64:
-		return int(t), nil
+		return int64(t), nil
 	case float32:
-		return int(t), nil
+		return int64(t), nil
 	case float64:
-		return int(t), nil
+		return int64(t), nil
 	case json.Number:
 		num, err := t.Int64()
-		return int(num), err
+		return int64(num), err
 	case string:
 		return strconv.Atoi(t)
 	}
